Ajout du module news

+++ app/src/main/java/com/pro/glagouy/calculatriceandroid/activities/MainActivity.java
@@ -12,8 +12,8 @@ import androidx.appcompat.app.AppCompatActivity;
 
 public class MainActivity extends AppCompatActivity {
     private String displayValue = "";
-    private Integer value1;
-    private Integer value2;
+    private Double value1;
+    private Double value2;
     private TextView textView;
     private String operator;
 
@@ -56,10 +56,10 @@ public class MainActivity extends AppCompatActivity {
      */
     public void calculate(View view) {
         String[] values = this.displayValue.split(String.format("\\%s", this.operator));
-        this.value1 = Integer.parseInt(values[0]);
-        this.value2 = Integer.parseInt(values[1]);
+        this.value1 = Double.parseDouble(values[0]);
+        this.value2 = Double.parseDouble(values[1]);
 
-        Integer result;
+        Double result;
         try {
             result = new Operation(this.value1, this.value2, operator).calculate();
             this.displayValue = result.toString();
